Patch for support ltalloc on SPARC architecture
Hi, Mario.
I've be forced to write this patch due to I require to use library on SPARC processors, and PAUSE definition is not compatible (syntaxically) with SPARC assembler. I've tried to write patch most compatible for most SPARC processors. It was tested with GCC 5.2 on SPARCv9 and seems ok. Library usable.